added the src/algorithm subdirectory and reorganized code.
