+++ include/planner.h
@@ -9,7 +9,7 @@
  * @copyright Copyright (c) 2024
  * 
  */
-#include "algorithm.h"
+#include "algorithm/algorithm.h"
 #include "network.h"
 #include "stop.h"
 
@@ -27,7 +27,6 @@ namespace supercharger
   class RoutePlanner
   {
     public:
-      // TODO: Change to accepting r-value reference?
       RoutePlanner(AlgoType&&, CostFcnType&&);
 
       // TODO: Add ctor that also takes max range and speed
@@ -45,8 +44,8 @@ namespace supercharger
 
     private:
       // TODO: Switch to unique_ptr?
-      Charger* origin_;
-      Charger* destination_;
+      Charger* origin_{nullptr};
+      Charger* destination_{nullptr};
 
       // Store some hard-coded constants
       const double max_range_{320};   // [km]
