+++ src/algorithm/brute_force.cpp
@@ -1,17 +1,15 @@
 /**
- * @file algorithm.cpp
+ * @file brute_force.cpp
  * @author your name (you@domain.com)
  * @brief 
  * @version 0.1
- * @date 2024-08-13
+ * @date 2024-08-23
  * 
  * @copyright Copyright (c) 2024
  * 
  */
-
-#include "algorithm.h"
+#include "algorithm/brute_force.h"
 #include "logging.h"
-#include "math.h"
 // Need to include planner.h here to avoid "pointer or reference to incomplete
 // type is not allowed" errors erlated to using the route_planner_ pointer.
 #include "planner.h"
@@ -19,53 +17,9 @@
 #include <algorithm>
 #include <map>
 
+
 namespace supercharger
 {
-  // NOTE: Do NOT repeat the 'static keyword at the cpp file level
-  // NOTE: I think this must be defined at the cpp file level because otherwise
-  // I get a "not declared in this scope error" related to the derived planning
-  // algorithm types (BruteForce, Dijkstra's, etc.). 
-  std::unique_ptr<PlanningAlgorithm> PlanningAlgorithm::GetPlanner(
-    AlgorithmType&& algo_type,
-    CostFunctionType&& cost_type = CostFunctionType::NONE)
-  {
-    switch ( algo_type )
-    {
-      case AlgorithmType::BRUTE_FORCE:
-        // NOTE: There is an important difference between public and protected
-        // inheritance when it comes to the compiler "being aware" that a
-        // unique_ptr of the base class (the return type of this function) can
-        // be initialized from unique_ptr of the derived class: public
-        // inheritance allows this, but protected inheritance does not. But why?
-        return std::make_unique<BruteForce>(cost_type);
-
-      case AlgorithmType::DIJKSTRAS:
-        // return std::make_unique<Dijstras>();
-        return std::unique_ptr<PlanningAlgorithm>(nullptr);
-
-      case AlgorithmType::ASTAR:
-        return std::unique_ptr<PlanningAlgorithm>(nullptr);
-        // return std::make_unique<AStar>();
-      
-      default:
-        return std::unique_ptr<PlanningAlgorithm>(nullptr);
-    }
-  }
-
-  /**
-   * @brief Effectively a wrapper around the great_circle_distance() function.
-   * 
-   * @param charger1
-   * @param charger2
-   * @return double 
-   */
-  double PlanningAlgorithm::ComputeDistance(
-    const Charger* const charger1, const Charger* const charger2) const
-  {
-    return great_circle_distance(
-      charger1->lat, charger1->lon, charger2->lat, charger2->lon);
-  }
-
   double BruteForce::ComputeChargeTime_(
     const Stop& current_stop, const Charger* const next_charger) const
   {
@@ -144,8 +98,7 @@ namespace supercharger
    * "reachable" and "closer to" sets. I beleive this will require defining
    * both operator== and a hash function for the supercharger::Charger class.
    * 
-   * @param current_stop 
-   * @return Stop 
+   * @param route 
    */
   void BruteForce::PlanRoute(std::vector<Stop>& route) {
     // Get the current stop
@@ -279,36 +232,4 @@ namespace supercharger
 
     return;
   }
-
-  void Dijkstras::PlanRoute(std::vector<Stop>& route) {
-    // 1. Mark all "nodes"/stops as unvisited
-
-    // 2. Assign a "distance to start" value to each node in the network.
-    // Initially, this value will be "infinity" since no path is yet known to
-    // these "unvisited" nodes.
-
-    // 3. From the unvisited set, select the current node to be the one with
-    // the smallest known distance from the origin. If the current node is the
-    // destination node, we're done; otherwise, continue to find the shortest
-    // path to all reachable nodes.
-
-    // 4. For the current node, consider all of its unvisited neighbors and
-    // update their distance through the current node. Compare the newly
-    // calculated distance to the one currently assigned to the neighbor, and
-    // assign the neighbor the smaller distance.
-
-    // 5. Once we've considered all the unvisited neighbors of the current node,
-    // mark the current node as visited and remove it from the unvisited set.
-
-    // 6. Return to step 3.
-
-    // 7. Once the loop (steps 3-5) exits, every node will contain the shortest
-    // distance from the start node.
-    return;
-  }
-
-  void AStar::PlanRoute(std::vector<Stop>& route) {
-    return;
-  }
-
 } // end namespace supercharger
